move all of extension setup in to setup class
