+++ extension/src/extension.ts
@@ -1,33 +1,21 @@
 import {
   commands,
   env,
-  Event,
   EventEmitter,
   ExtensionContext,
   ViewColumn
 } from 'vscode'
-import isEmpty from 'lodash.isempty'
 import { DvcExecutor } from './cli/dvc/executor'
 import { DvcRunner } from './cli/dvc/runner'
 import { DvcReader } from './cli/dvc/reader'
 import { Config } from './config'
 import { Context } from './context'
-import { isPythonExtensionInstalled } from './extensions/python'
 import { WorkspaceExperiments } from './experiments/workspace'
 import { registerExperimentCommands } from './experiments/commands/register'
 import { registerPlotsCommands } from './plots/commands/register'
-import {
-  findAbsoluteDvcRootPath,
-  findDvcRootPaths,
-  findSubRootPaths
-} from './fileSystem'
 import { RepositoriesTree } from './repository/model/tree'
-import { IExtension } from './interfaces'
 import { registerRepositoryCommands } from './repository/commands/register'
 import { ResourceLocator } from './resourceLocator'
-import { definedAndNonEmpty } from './util/array'
-import { setup, setupWithGlobalRecheck, setupWorkspace } from './setup'
-import { Status } from './status'
 import { reRegisterVsCodeCommands } from './vscode/commands'
 import { InternalCommands } from './commands/internal'
 import { ExperimentsColumnsTree } from './experiments/columns/tree'
@@ -36,22 +24,15 @@ import { ExperimentsTree } from './experiments/model/tree'
 import { ExperimentsFilterByTree } from './experiments/model/filterBy/tree'
 import {
   Context as VsCodeContext,
-  getDvcRootFromContext,
-  setContextValue
+  getDvcRootFromContext
 } from './vscode/context'
 import { OutputChannel } from './vscode/outputChannel'
-import {
-  getFirstWorkspaceFolder,
-  getWorkspaceFolderCount,
-  getWorkspaceFolders
-} from './vscode/workspaceFolders'
 import {
   getTelemetryReporter,
   sendTelemetryEvent,
   sendTelemetryEventAndThrow
 } from './telemetry'
-import { EventName } from './telemetry/constants'
-import { RegisteredCliCommands, RegisteredCommands } from './commands/external'
+import { RegisteredCommands } from './commands/external'
 import { StopWatch } from './util/time'
 import {
   registerWalkthroughCommands,
@@ -63,17 +44,14 @@ import { WorkspacePlots } from './plots/workspace'
 import { PlotsPathsTree } from './plots/paths/tree'
 import { Disposable } from './class/dispose'
 import { collectWorkspaceScale } from './telemetry/collect'
-import { createFileSystemWatcher } from './fileSystem/watcher'
 import { GitExecutor } from './cli/git/executor'
 import { GitReader } from './cli/git/reader'
 import { Setup } from './setup/index'
 
-export class Extension extends Disposable implements IExtension {
+export class Extension extends Disposable {
   protected readonly internalCommands: InternalCommands
 
   private readonly resourceLocator: ResourceLocator
-  private readonly config: Config
-  private dvcRoots: string[] = []
   private readonly repositories: WorkspaceRepositories
   private readonly experiments: WorkspaceExperiments
   private readonly plots: WorkspacePlots
@@ -84,20 +62,10 @@ export class Extension extends Disposable implements IExtension {
   private readonly dvcRunner: DvcRunner
   private readonly gitExecutor: GitExecutor
   private readonly gitReader: GitReader
-  private readonly status: Status
-  private cliAccessible = false
-  private cliCompatible: boolean | undefined
-
-  private readonly workspaceChanged: EventEmitter<void> = this.dispose.track(
-    new EventEmitter()
-  )
 
   private updatesPaused: EventEmitter<boolean> = this.dispose.track(
     new EventEmitter<boolean>()
   )
-
-  private readonly onDidChangeWorkspace: Event<void> =
-    this.workspaceChanged.event
 
   constructor(context: ExtensionContext) {
     super()
@@ -106,18 +74,15 @@ export class Extension extends Disposable implements IExtension {
 
     this.dispose.track(getTelemetryReporter())
 
-    this.setCommandsAvailability(false)
-    this.setProjectAvailability()
-
     this.resourceLocator = this.dispose.track(
       new ResourceLocator(context.extensionUri)
     )
 
-    this.config = this.dispose.track(new Config())
+    const config = this.dispose.track(new Config())
 
-    this.dvcExecutor = this.dispose.track(new DvcExecutor(this.config))
-    this.dvcReader = this.dispose.track(new DvcReader(this.config))
-    this.dvcRunner = this.dispose.track(new DvcRunner(this.config))
+    this.dvcExecutor = this.dispose.track(new DvcExecutor(config))
+    this.dvcReader = this.dispose.track(new DvcReader(config))
+    this.dvcRunner = this.dispose.track(new DvcRunner(config))
 
     this.gitExecutor = this.dispose.track(new GitExecutor())
     this.gitReader = this.dispose.track(new GitReader())
@@ -138,17 +103,6 @@ export class Extension extends Disposable implements IExtension {
       new InternalCommands(outputChannel, ...clis)
     )
 
-    this.status = this.dispose.track(
-      new Status(
-        this.config,
-        this.dvcExecutor,
-        this.dvcReader,
-        this.dvcRunner,
-        this.gitExecutor,
-        this.gitReader
-      )
-    )
-
     this.experiments = this.dispose.track(
       new WorkspaceExperiments(
         this.internalCommands,
@@ -157,31 +111,8 @@ export class Extension extends Disposable implements IExtension {
       )
     )
 
-    const onDidChangeHasData = this.experiments.columnsChanged.event
-    this.dispose.track(
-      onDidChangeHasData(() => {
-        this.changeSetupStep()
-        setContextValue('dvc.project.hasData', this.experiments.getHasData())
-      })
-    )
-
     this.plots = this.dispose.track(
       new WorkspacePlots(this.internalCommands, context.workspaceState)
-    )
-
-    this.setup = this.dispose.track(
-      new Setup(
-        '',
-        this.resourceLocator.dvcIcon,
-        () => this.initializeDvc(),
-        () => this.initializeGit(),
-        () => this.showExperiments(this.dvcRoots[0]),
-        () => this.getCliCompatible(),
-        () => this.needsGitInit(),
-        needsGitInit => this.canGitInitialize(needsGitInit),
-        () => this.hasRoots(),
-        () => this.experiments.getHasData()
-      )
     )
 
     this.repositories = this.dispose.track(
@@ -222,50 +153,28 @@ export class Extension extends Disposable implements IExtension {
       new RepositoriesTree(this.internalCommands, this.repositories)
     )
 
-    setupWithGlobalRecheck(this)
-      .then(async () => {
-        sendTelemetryEvent(
-          EventName.EXTENSION_LOAD,
-          await this.getEventProperties(),
-          { duration: stopWatch.getElapsedTime() }
-        )
-      })
-      .catch(async error =>
-        sendTelemetryEventAndThrow(
-          EventName.EXTENSION_LOAD,
-          error,
-          stopWatch.getElapsedTime(),
-          await this.getEventProperties()
-        )
+    this.setup = this.dispose.track(
+      new Setup(
+        stopWatch,
+        config,
+        this.dvcExecutor,
+        this.dvcReader,
+        this.dvcRunner,
+        this.gitExecutor,
+        this.gitReader,
+        () => this.initialize(),
+        () => this.resetMembers(),
+        this.experiments,
+        this.internalCommands,
+        this.resourceLocator.dvcIcon,
+        () =>
+          collectWorkspaceScale(
+            this.getRoots(),
+            this.experiments,
+            this.plots,
+            this.repositories
+          )
       )
-
-    this.dispose.track(
-      this.onDidChangeWorkspace(() => {
-        setup(this)
-      })
-    )
-
-    this.dispose.track(
-      this.config.onDidChangeExecutionDetails(async () => {
-        const stopWatch = new StopWatch()
-        try {
-          this.changeSetupStep()
-          await setup(this)
-
-          return sendTelemetryEvent(
-            EventName.EXTENSION_EXECUTION_DETAILS_CHANGED,
-            await this.getEventProperties(),
-            { duration: stopWatch.getElapsedTime() }
-          )
-        } catch (error: unknown) {
-          return sendTelemetryEventAndThrow(
-            EventName.EXTENSION_EXECUTION_DETAILS_CHANGED,
-            error as Error,
-            stopWatch.getElapsedTime(),
-            await this.getEventProperties()
-          )
-        }
-      })
     )
 
     registerExperimentCommands(this.experiments, this.internalCommands)
@@ -313,18 +222,8 @@ export class Extension extends Disposable implements IExtension {
     )
 
     this.internalCommands.registerExternalCommand(
-      RegisteredCommands.EXTENSION_CHECK_CLI_COMPATIBLE,
-      () => setup(this)
-    )
-
-    this.internalCommands.registerExternalCommand(
       RegisteredCommands.EXTENSION_SHOW_OUTPUT,
       () => outputChannel.show()
-    )
-
-    this.internalCommands.registerExternalCliCommand(
-      RegisteredCliCommands.INIT,
-      () => this.initializeDvc()
     )
 
     registerRepositoryCommands(this.repositories, this.internalCommands)
@@ -336,73 +235,28 @@ export class Extension extends Disposable implements IExtension {
       context.extension.packageJSON.contributes.walkthroughs[0].id
     )
 
-    this.dispose.track(
-      commands.registerCommand(
-        RegisteredCommands.EXTENSION_SETUP_WORKSPACE,
-        () => this.setupWorkspace()
-      )
-    )
-
     showWalkthroughOnFirstUse(env.isNewAppInstall)
     this.dispose.track(recommendRedHatExtensionOnce())
-
-    this.watchForVenvChanges()
-  }
-
-  public async showSetup() {
-    return await this.setup.showWebview()
-  }
-
-  public shouldWarnUserIfCLIUnavailable() {
-    return this.hasRoots() && !this.setup.isFocused()
-  }
-
-  public async isPythonExtensionUsed() {
-    await this.config.isReady()
-    return !!this.config.isPythonExtensionUsed()
-  }
-
-  public async getCliVersion(cwd: string, tryGlobalCli?: true) {
-    await this.config.isReady()
-    try {
-      return await this.dvcReader.version(cwd, tryGlobalCli)
-    } catch {}
-  }
-
-  public unsetPythonBinPath() {
-    this.config.unsetPythonBinPath()
-  }
-
-  public async setRoots() {
-    const nestedRoots = await Promise.all(
-      getWorkspaceFolders().map(workspaceFolder =>
-        this.findDvcRoots(workspaceFolder)
-      )
-    )
-    this.dvcRoots = nestedRoots.flat().sort()
-
-    this.changeSetupStep()
-    return this.setProjectAvailability()
   }
 
   public getRoots() {
-    return this.dvcRoots
+    return this.setup.getRoots()
   }
 
   public async initialize() {
     this.resetMembers()
 
     await Promise.all([
-      this.repositories.create(this.dvcRoots, this.updatesPaused),
-      this.repositoriesTree.initialize(this.dvcRoots),
+      this.repositories.create(this.getRoots(), this.updatesPaused),
+      this.repositoriesTree.initialize(this.getRoots()),
       this.experiments.create(
-        this.dvcRoots,
+        this.getRoots(),
         this.updatesPaused,
         this.resourceLocator
       )
     ])
     this.plots.create(
-      this.dvcRoots,
+      this.getRoots(),
       this.updatesPaused,
       this.resourceLocator,
       this.experiments
@@ -415,10 +269,6 @@ export class Extension extends Disposable implements IExtension {
     ])
   }
 
-  public hasRoots() {
-    return definedAndNonEmpty(this.dvcRoots)
-  }
-
   public resetMembers() {
     this.repositories.reset()
     this.repositoriesTree.initialize([])
@@ -426,176 +276,8 @@ export class Extension extends Disposable implements IExtension {
     this.plots.reset()
   }
 
-  public setCliCompatible(compatible: boolean | undefined) {
-    this.cliCompatible = compatible
-    const incompatible = compatible === undefined ? undefined : !compatible
-    setContextValue('dvc.cli.incompatible', incompatible)
-  }
-
-  public setAvailable(available: boolean) {
-    this.status.setAvailability(available)
-    this.setCommandsAvailability(available)
-    this.cliAccessible = available
-    this.changeSetupStep()
-    return available
-  }
-
-  public getAvailable() {
-    return this.cliAccessible
-  }
-
-  public async initializeDvc() {
-    const root = getFirstWorkspaceFolder()
-    if (root) {
-      await this.dvcExecutor.init(root)
-      this.workspaceChanged.fire()
-    }
-  }
-
   private async showExperiments(dvcRoot: string) {
     return await this.experiments.showWebview(dvcRoot, ViewColumn.Active)
-  }
-
-  private setCommandsAvailability(available: boolean) {
-    setContextValue('dvc.commands.available', available)
-  }
-
-  private async setupWorkspace() {
-    const stopWatch = new StopWatch()
-    try {
-      const previousCliPath = this.config.getCliPath()
-      const previousPythonPath = this.config.getPythonBinPath()
-
-      const completed = await setupWorkspace(() =>
-        this.config.setPythonAndNotifyIfChanged()
-      )
-      sendTelemetryEvent(
-        RegisteredCommands.EXTENSION_SETUP_WORKSPACE,
-        { completed },
-        {
-          duration: stopWatch.getElapsedTime()
-        }
-      )
-
-      const executionDetailsUnchanged =
-        this.config.getCliPath() === previousPythonPath &&
-        this.config.getPythonBinPath() === previousCliPath
-
-      if (completed && !this.cliAccessible && executionDetailsUnchanged) {
-        this.workspaceChanged.fire()
-      }
-
-      return completed
-    } catch (error: unknown) {
-      return sendTelemetryEventAndThrow(
-        RegisteredCommands.EXTENSION_SETUP_WORKSPACE,
-        error as Error,
-        stopWatch.getElapsedTime()
-      )
-    }
-  }
-
-  private setProjectAvailability() {
-    const available = this.hasRoots()
-    setContextValue('dvc.project.available', available)
-  }
-
-  private async findDvcRoots(cwd: string): Promise<string[]> {
-    const dvcRoots = await findDvcRootPaths(cwd)
-    if (definedAndNonEmpty(dvcRoots)) {
-      return dvcRoots
-    }
-
-    await this.config.isReady()
-    return findAbsoluteDvcRootPath(cwd, this.dvcReader.root(cwd))
-  }
-
-  private async getEventProperties() {
-    return {
-      ...(this.cliAccessible
-        ? await collectWorkspaceScale(
-            this.dvcRoots,
-            this.experiments,
-            this.plots,
-            this.repositories
-          )
-        : {}),
-      cliAccessible: this.cliAccessible,
-      dvcPathUsed: !!this.config.getCliPath(),
-      dvcRootCount: this.dvcRoots.length,
-      msPythonInstalled: isPythonExtensionInstalled(),
-      msPythonUsed: this.config.isPythonExtensionUsed(),
-      pythonPathUsed: !!this.config.getPythonBinPath(),
-      workspaceFolderCount: getWorkspaceFolderCount()
-    }
-  }
-
-  private watchForVenvChanges() {
-    return createFileSystemWatcher(
-      disposable => this.dispose.track(disposable),
-      '**/dvc{,.exe}',
-      async path => {
-        if (!path) {
-          return
-        }
-
-        const previousPythonBinPath = this.config.getPythonBinPath()
-        await this.config.setPythonBinPath()
-
-        const trySetupWithVenv =
-          previousPythonBinPath !== this.config.getPythonBinPath()
-
-        if (!this.cliAccessible || !this.cliCompatible || trySetupWithVenv) {
-          setup(this)
-        }
-      }
-    )
-  }
-
-  private getCliCompatible() {
-    return this.cliCompatible
-  }
-
-  private initializeGit() {
-    const cwd = getFirstWorkspaceFolder()
-    if (cwd) {
-      this.gitExecutor.init(cwd)
-      this.workspaceChanged.fire()
-    }
-  }
-
-  private async canGitInitialize(needsGitInit: boolean) {
-    if (!needsGitInit) {
-      return false
-    }
-    const nestedRoots = await Promise.all(
-      getWorkspaceFolders().map(workspaceFolder =>
-        findSubRootPaths(workspaceFolder, '.git')
-      )
-    )
-
-    return isEmpty(nestedRoots.flat())
-  }
-
-  private async needsGitInit() {
-    if (this.hasRoots()) {
-      return false
-    }
-
-    const cwd = getFirstWorkspaceFolder()
-    if (!cwd) {
-      return undefined
-    }
-
-    try {
-      return !(await this.gitExecutor.getGitRepositoryRoot(cwd))
-    } catch {
-      return true
-    }
-  }
-
-  private changeSetupStep() {
-    this.setup.sendDataToWebview()
   }
 }
