+++ extension/src/test/suite/setup/index.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, it, suite } from 'mocha'
 import { expect } from 'chai'
-import { restore, spy } from 'sinon'
+import { restore, spy, stub } from 'sinon'
 import { buildSetup } from './util'
 import { closeAllEditors, getMessageReceivedEmitter } from '../util'
 import { WEBVIEW_TEST_TIMEOUT } from '../timeouts'
@@ -25,10 +25,7 @@ suite('Setup Test Suite', () => {
 
   describe('Setup', () => {
     it('should handle an initialize git message from the webview', async () => {
-      const { messageSpy, setup, mockInitializeGit } = buildSetup(
-        disposable,
-        true
-      )
+      const { messageSpy, setup, mockInitializeGit } = buildSetup(disposable)
 
       const webview = await setup.showWebview()
       await webview.isReady()
@@ -44,10 +41,7 @@ suite('Setup Test Suite', () => {
     }).timeout(WEBVIEW_TEST_TIMEOUT)
 
     it('should handle an initialize project message from the webview', async () => {
-      const { messageSpy, setup, mockInitializeDvc } = buildSetup(
-        disposable,
-        true
-      )
+      const { messageSpy, setup, mockInitializeDvc } = buildSetup(disposable)
 
       const webview = await setup.showWebview()
       await webview.isReady()
@@ -134,17 +128,16 @@ suite('Setup Test Suite', () => {
     }).timeout(WEBVIEW_TEST_TIMEOUT)
 
     it('should close the webview and open the experiments when the setup is done', async () => {
-      const { setup, mockOpenExperiments } = buildSetup(
-        disposable,
-        true,
-        true,
-        true
-      )
+      const { setup, mockOpenExperiments } = buildSetup(disposable, true)
 
       const closeWebviewSpy = spy(BaseWebview.prototype, 'dispose')
 
       const webview = await setup.showWebview()
       await webview.isReady()
+
+      stub(setup, 'hasRoots').returns(true)
+      setup.setCliCompatible(true)
+      setup.setAvailable(true)
 
       expect(closeWebviewSpy).to.be.calledOnce
       expect(mockOpenExperiments).to.be.calledOnce
